avances

+++ TestTorneo/app/Services/TorneoService.php
@@ -9,77 +9,83 @@ use App\Models\Torneo;
 use App\Models\Partida;
 use App\Interfaces\ITorneoService;
 use App\Interfaces\IGanadorStrategy;
+use App\Interfaces\IJugadorService;
+use Exception;
 use Illuminate\Support\Facades\DB;
 
 class TorneoService implements ITorneoService
 {
     private ITorneoRepository $torneoRepository;
     private IPartidaRepository $partidaRepository;
-    private IGanadorStrategy $ganadorStrategy;
+    private IJugadorService $jugadorService;
 
     public function __construct(
         ITorneoRepository $torneoRepository,
         IPartidaRepository $partidaRepository,
-        IGanadorStrategy $ganadorStrategy
+        IJugadorService $jugadorService,
     ) {
         $this->torneoRepository = $torneoRepository;
         $this->partidaRepository = $partidaRepository;
-        $this->ganadorStrategy = $ganadorStrategy;
+        $this->jugadorService = $jugadorService;
     }
 
-
-    public function crearPartida(Torneo $torneo, Jugador $jugador1, Jugador $jugador2, int $ronda): Partida
-    {
-        return $this->partidaRepository->create([
-            'torneo_id' => $torneo->id,
-            'jugador1_id' => $jugador1->id,
-            'jugador1_type' => get_class($jugador1),
-            'jugador2_id' => $jugador2->id,
-            'jugador2_type' => get_class($jugador2),
-            'ronda' => $ronda
-        ]);
-    }
-
-    public function obtenerTorneos()
+    public function getAll()
     {
         return $this->torneoRepository->getAll();
     }
 
-    public function crearTorneo(array $data): Torneo
+    public function create(array $data): Torneo
     {
+        //VALIDAR
         return $this->torneoRepository->create($data);
     }
 
-    public function eliminarTorneo(int $id): bool
+    public function delete(int $id): bool
     {
+        //Validar que exista el torneo
+
+        $torneo = $this->torneoRepository->findById($id);
+
+        if (!$torneo) {
+            return false;
+        }
+
+        // Eliminar partidas del torneo antes de eliminar el torneo
+        $this->partidaRepository->deleteByTorneoId($id);
+
         return $this->torneoRepository->delete($id);
     }
 
-
-    public function comenzarTorneo(int $torneoId): ?Torneo
+    public function restore(int $id): bool
     {
-        return new Torneo();
+        $jugador = $this->torneoRepository->findByIdWithTrashed($id);
+
+        if (!$jugador) {
+            throw new Exception("Torneo no encontrado o no está eliminado.");
+        }
+
+        return $this->torneoRepository->restore($id);
     }
 
     public function determinarGanador(Partida $partida)
     {
-        $ganador = $this->ganadorStrategy->determinarGanador(
-            $partida->jugador1,
-            $partida->jugador2
-        );
+
+        $ganador = $this->jugadorService->calcularPuntaje($partida->jugador1) >= $this->jugadorService->calcularPuntaje($partida->jugador2) ? $partida->jugador1 : $partida->jugador2;
 
         $this->partidaRepository->update($partida->id, ['ganador_id' => $ganador->id]);
 
         return $ganador;
     }
 
-    private function jugarPartidas(Torneo $torneo, $jugadores): Torneo
+    public function actualizarGanador(int $torneoId, int $ganadorId): bool
     {
-        return new Torneo();
-    }
+        $torneo = $this->torneoRepository->findById($torneoId);
 
-    public function totalPartidasJugadas(Torneo $torneo): int
-    {
-        return $torneo->partidas()->count();
+        if (!$torneo) {
+            throw new Exception("Torneo no encontrado.");
+        }
+
+        // Actualizamos el torneo con el ganador
+        return $this->torneoRepository->update($torneoId, ['ganador_id' => $ganadorId]);
     }
 }
